Extract portfolio holdings contracts
Move shared portfolio identity/summary and holdings/book response contracts into focused modules while preserving app.contracts.portfolio compatibility re-exports. Update internal holdings consumers to import the narrower modules directly and refresh the governed monetary-float allowlist for moved baseline fields.

Evidence: python -m ruff check touched files passed; python -m ruff format --check touched files passed; python -m mypy touched source files passed for 12 source files; python scripts/check_monetary_float_usage.py passed with Findings=159 allowlisted=159; focused pytest portfolio holdings/contract/workflow/workspace tests passed with 24 tests.
