+++ src/app/routers/portfolio_book.py
@@ -1,8 +1,6 @@
 from fastapi import APIRouter, Query
 
-from app.contracts.portfolio import (
-    PortfolioBookResponse,
-)
+from app.contracts.portfolio_holdings import PortfolioBookResponse
 from app.middleware.correlation import correlation_id_var
 from app.services.portfolio_service_provider import portfolio_service
