+++ src/app/routers/portfolio_positions.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Query
 
-from app.contracts.portfolio import PortfolioPositionBookResponse
+from app.contracts.portfolio_holdings import PortfolioPositionBookResponse
 from app.middleware.correlation import correlation_id_var
 from app.services.portfolio_service_provider import portfolio_service
