+++ src/app/services/portfolio_workflow.py
@@ -1,12 +1,11 @@
 from dataclasses import dataclass
 
 from app.contracts.portfolio import (
-    PortfolioAllocationView,
     PortfolioOperationalReadiness,
-    PortfolioPositionView,
-    PortfolioSummary,
     PortfolioWorkspaceResponse,
 )
+from app.contracts.portfolio_core import PortfolioSummary
+from app.contracts.portfolio_holdings import PortfolioAllocationView, PortfolioPositionView
 from app.contracts.portfolio_workflow import (
     PortfolioReadinessIndicator,
     PortfolioWorkflowAction,
